Inform users about network and server timeout errors.
Timeout or network errors previously appeared to the user as "a flicker of
spinner followed by no results or an empty modal", leaving them to figure
out for themselves what exactly happened. Not good.

Detecting timeout errors is tricky. We can be sure that the response will be
empty if server timed-out. But empty response can be caused due to myriad of
network issues.

So we check for empty response and notify users that either the server timed
out (and hence she should contact admin), or there's a temporary network issue
(and hence she should try again later). As a laster resort, the user should
seek help on the forum.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>